refactor: add group selection functionality and save selected group to JSON
